+++ rend3/src/renderer/render.rs
@@ -1,13 +1,8 @@
 use crate::{
-    format_sso, instruction::Instruction, resources::InternalTexture, util::typedefs::RendererStatistics,
-    ManagerReadyOutput, RenderRoutine, Renderer,
+    instruction::Instruction, util::typedefs::RendererStatistics, ManagerReadyOutput, RenderRoutine, Renderer,
 };
-use rend3_types::{MipmapCount, MipmapSource};
-use std::{num::NonZeroU32, sync::Arc};
-use wgpu::{
-    util::DeviceExt, CommandEncoderDescriptor, Extent3d, ImageCopyTexture, ImageDataLayout, Origin3d, TextureAspect,
-    TextureDescriptor, TextureDimension, TextureUsages, TextureViewDescriptor, TextureViewDimension,
-};
+use std::sync::Arc;
+use wgpu::CommandEncoderDescriptor;
 
 pub fn render_loop<Input, Output>(
     renderer: Arc<Renderer>,
@@ -21,6 +16,9 @@ pub fn render_loop<Input, Output>(
 
     let mut instructions = renderer.instructions.consumer.lock();
 
+    // 16 encoders is a reasonable default
+    let mut cmd_bufs = Vec::with_capacity(16);
+
     let mut encoder = renderer.device.create_command_encoder(&CommandEncoderDescriptor {
         label: Some("primary encoder"),
     });
@@ -33,7 +31,7 @@ pub fn render_loop<Input, Output>(
     let mut directional_light_manager = renderer.directional_light_manager.write();
     let mut camera_manager = renderer.camera_manager.write();
     let mut profiler = renderer.profiler.lock();
-    let mut mipmap_generator = renderer.mipmap_generator.lock();
+
     {
         profiling::scope!("Instruction Processing");
         for cmd in instructions.drain(..) {
@@ -44,167 +42,20 @@ pub fn render_loop<Input, Output>(
                     mesh_manager.fill(&renderer.device, &renderer.queue, &mut encoder, &handle, mesh);
                     profiler.end_scope(&mut encoder);
                 }
-                Instruction::AddTexture2D { handle, texture } => {
-                    profiling::scope!("Add Texture 2D");
-                    let size = Extent3d {
-                        width: texture.size.x,
-                        height: texture.size.y,
-                        depth_or_array_layers: 1,
-                    };
-
-                    let mip_level_count = match texture.mip_count {
-                        MipmapCount::Specific(v) => v.get(),
-                        MipmapCount::Maximum => size.max_mips(),
-                    };
-
-                    let desc = TextureDescriptor {
-                        label: None,
-                        size,
-                        mip_level_count,
-                        sample_count: 1,
-                        dimension: TextureDimension::D2,
-                        format: texture.format,
-                        usage: TextureUsages::TEXTURE_BINDING | TextureUsages::COPY_SRC | TextureUsages::COPY_DST,
-                    };
-
-                    let uploaded_tex = match texture.mip_source {
-                        MipmapSource::Uploaded => {
-                            renderer
-                                .device
-                                .create_texture_with_data(&renderer.queue, &desc, &texture.data)
-                        }
-                        MipmapSource::Generated => {
-                            let desc = TextureDescriptor {
-                                usage: desc.usage | TextureUsages::RENDER_ATTACHMENT,
-                                ..desc
-                            };
-                            let tex = renderer.device.create_texture(&desc);
-
-                            let format_desc = texture.format.describe();
-
-                            // write first level
-                            renderer.queue.write_texture(
-                                ImageCopyTexture {
-                                    texture: &tex,
-                                    mip_level: 0,
-                                    origin: Origin3d::ZERO,
-                                    aspect: TextureAspect::All,
-                                },
-                                &texture.data,
-                                ImageDataLayout {
-                                    offset: 0,
-                                    bytes_per_row: NonZeroU32::new(
-                                        format_desc.block_size as u32
-                                            * (size.width / format_desc.block_dimensions.0 as u32),
-                                    ),
-                                    rows_per_image: None,
-                                },
-                                size,
-                            );
-
-                            // generate mipmaps
-                            mipmap_generator.generate_mipmaps(
-                                &renderer.device,
-                                &mut profiler,
-                                &mut encoder,
-                                &tex,
-                                &desc,
-                            );
-
-                            tex
-                        }
-                    };
-
-                    let view = uploaded_tex.create_view(&TextureViewDescriptor::default());
-
-                    texture_manager_2d.fill(&handle, desc, uploaded_tex, view);
-                }
-                Instruction::AddTexture2DFromTexture { handle, texture } => {
-                    profiling::scope!("Add Texture 2D From Texture");
-                    profiler.begin_scope("Add Texture 2D From Texture", &mut encoder, &renderer.device);
-                    let InternalTexture {
-                        texture: old_texture,
-                        desc: old_texture_desc,
-                    } = texture_manager_2d.get_internal(texture.src.get_raw());
-
-                    let new_size = old_texture_desc.mip_level_size(texture.start_mip).unwrap();
-
-                    let mip_level_count = texture
-                        .mip_count
-                        .map_or_else(|| old_texture_desc.mip_level_count - texture.start_mip, |c| c.get());
-
-                    let desc = TextureDescriptor {
-                        size: new_size,
-                        mip_level_count,
-                        ..old_texture_desc.clone()
-                    };
-
-                    let tex = renderer.device.create_texture(&desc);
-
-                    let view = tex.create_view(&TextureViewDescriptor::default());
-
-                    for new_mip in 0..mip_level_count {
-                        let old_mip = new_mip + texture.start_mip;
-
-                        let label = format_sso!("mip {} to {}", old_mip, new_mip);
-                        profiling::scope!(&label);
-                        profiler.begin_scope(&label, &mut encoder, &renderer.device);
-
-                        encoder.copy_texture_to_texture(
-                            ImageCopyTexture {
-                                texture: old_texture,
-                                mip_level: old_mip,
-                                origin: Origin3d::ZERO,
-                                aspect: TextureAspect::All,
-                            },
-                            ImageCopyTexture {
-                                texture: &tex,
-                                mip_level: new_mip,
-                                origin: Origin3d::ZERO,
-                                aspect: TextureAspect::All,
-                            },
-                            old_texture_desc.mip_level_size(old_mip).unwrap(),
-                        );
-
-                        profiler.end_scope(&mut encoder);
+                Instruction::AddTexture {
+                    handle,
+                    desc,
+                    texture,
+                    view,
+                    buffer,
+                    cube,
+                } => {
+                    cmd_bufs.extend(buffer);
+                    if cube {
+                        texture_manager_cube.fill(&handle, desc, texture, view);
+                    } else {
+                        texture_manager_2d.fill(&handle, desc, texture, view);
                     }
-                    profiler.end_scope(&mut encoder);
-
-                    texture_manager_2d.fill(&handle, desc, tex, view);
-                }
-                Instruction::AddTextureCube { handle, texture } => {
-                    profiling::scope!("Add Texture Cube");
-                    let size = Extent3d {
-                        width: texture.size.x,
-                        height: texture.size.y,
-                        depth_or_array_layers: 6,
-                    };
-
-                    let mip_level_count = match texture.mip_count {
-                        MipmapCount::Specific(v) => v.get(),
-                        MipmapCount::Maximum => size.max_mips(),
-                    };
-
-                    let desc = TextureDescriptor {
-                        label: None,
-                        size,
-                        mip_level_count,
-                        sample_count: 1,
-                        dimension: TextureDimension::D2,
-                        format: texture.format,
-                        usage: TextureUsages::TEXTURE_BINDING | TextureUsages::COPY_DST,
-                    };
-
-                    let uploaded_tex = renderer
-                        .device
-                        .create_texture_with_data(&renderer.queue, &desc, &texture.data);
-
-                    let view = uploaded_tex.create_view(&TextureViewDescriptor {
-                        dimension: Some(TextureViewDimension::Cube),
-                        ..TextureViewDescriptor::default()
-                    });
-
-                    texture_manager_cube.fill(&handle, desc, uploaded_tex, view);
                 }
                 Instruction::AddMaterial { handle, material } => {
                     profiling::scope!("Add Texture Material");
@@ -267,12 +118,9 @@ pub fn render_loop<Input, Output>(
         material_manager,
         object_manager,
         directional_light_manager,
-        mipmap_generator,
         profiler,
     ));
 
-    // 16 encoders is a reasonable default
-    let mut cmd_bufs = Vec::with_capacity(16);
     let (sender, reciever) = flume::unbounded();
     cmd_bufs.push(encoder.finish());
